+++ internal/pkg/instrumentation/bpf/github.com/segmentio/kafka-go/producer/bpf_arm64_bpfel.go
@@ -16,11 +16,10 @@ type bpfKafkaRequestT struct {
 	StartTime uint64
 	EndTime   uint64
 	Psc       bpfSpanContext
-	TraceID   [16]uint8
 	Msgs      [10]struct {
-		SpanID [8]uint8
-		Topic  [256]int8
-		Key    [256]int8
+		Sc    bpfSpanContext
+		Topic [256]int8
+		Key   [256]int8
 	}
 	GlobalTopic   [256]int8
 	ValidMessages uint64
@@ -29,8 +28,10 @@ type bpfKafkaRequestT struct {
 type bpfSliceArrayBuff struct{ Buff [1024]uint8 }
 
 type bpfSpanContext struct {
-	TraceID [16]uint8
-	SpanID  [8]uint8
+	TraceID    [16]uint8
+	SpanID     [8]uint8
+	TraceFlags uint8
+	Padding    [7]uint8
 }
 
 // loadBpf returns the embedded CollectionSpec for bpf.