+++ internal/pkg/instrumentation/bpf/go.opentelemetry.io/otel/traceglobal/probe.go
@@ -1,30 +1,20 @@
 // Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// SPDX-License-Identifier: Apache-2.0
 
 package global
 
 import (
 	"encoding/binary"
+	"fmt"
 	"math"
-	"os"
 
+	"go.opentelemetry.io/auto/internal/pkg/inject"
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/probe"
+	"go.opentelemetry.io/auto/internal/pkg/process"
 	"go.opentelemetry.io/auto/internal/pkg/structfield"
 
-	"github.com/cilium/ebpf/link"
-	"github.com/cilium/ebpf/perf"
 	"github.com/go-logr/logr"
+	"github.com/hashicorp/go-version"
 	"golang.org/x/sys/unix"
 
 	"go.opentelemetry.io/otel/attribute"
@@ -32,7 +22,6 @@ import (
 	"go.opentelemetry.io/otel/trace"
 
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/context"
-	"go.opentelemetry.io/auto/internal/pkg/process"
 )
 
 //go:generate go run github.com/cilium/ebpf/cmd/bpf2go -target amd64,arm64 -cc clang -cflags $CFLAGS bpf ./bpf/probe.bpf.c
@@ -94,138 +83,70 @@ func New(logger logr.Logger) probe.Probe {
 				Key: "tracer_delegate_pos",
 				Val: structfield.NewID("go.opentelemetry.io/otel", "go.opentelemetry.io/otel/internal/global", "tracer", "delegate"),
 			},
+			probe.StructFieldConst{
+				Key: "tracer_name_pos",
+				Val: structfield.NewID("go.opentelemetry.io/otel", "go.opentelemetry.io/otel/internal/global", "tracer", "name"),
+			},
+			probe.StructFieldConst{
+				Key: "tracer_provider_pos",
+				Val: structfield.NewID("go.opentelemetry.io/otel", "go.opentelemetry.io/otel/internal/global", "tracer", "provider"),
+			},
+			probe.StructFieldConst{
+				Key: "tracer_provider_tracers_pos",
+				Val: structfield.NewID("go.opentelemetry.io/otel", "go.opentelemetry.io/otel/internal/global", "tracerProvider", "tracers"),
+			},
+			probe.StructFieldConst{
+				Key: "buckets_ptr_pos",
+				Val: structfield.NewID("std", "runtime", "hmap", "buckets"),
+			},
+			tracerIDContainsSchemaURL{},
 		},
-		Uprobes: []probe.Uprobe[bpfObjects]{
+		Uprobes: []probe.Uprobe{
 			{
-				Sym: "go.opentelemetry.io/otel/internal/global.(*tracer).Start",
-				Fn:  uprobeTracerStart,
+				Sym:         "go.opentelemetry.io/otel/internal/global.(*tracer).Start",
+				EntryProbe:  "uprobe_Start",
+				ReturnProbe: "uprobe_Start_Returns",
 			},
 			{
-				Sym: "go.opentelemetry.io/otel/internal/global.(*nonRecordingSpan).End",
-				Fn:  uprobeSpanEnd,
+				Sym:        "go.opentelemetry.io/otel/internal/global.(*nonRecordingSpan).End",
+				EntryProbe: "uprobe_End",
 			},
 			{
-				Sym:      "go.opentelemetry.io/otel/internal/global.(*nonRecordingSpan).SetAttributes",
-				Fn:       uprobeSetAttributes,
-				Optional: true,
+				Sym:        "go.opentelemetry.io/otel/internal/global.(*nonRecordingSpan).SetAttributes",
+				EntryProbe: "uprobe_SetAttributes",
+				Optional:   true,
 			},
 			{
-				Sym:      "go.opentelemetry.io/otel/internal/global.(*nonRecordingSpan).SetStatus",
-				Fn:       uprobeSetStatus,
-				Optional: true,
+				Sym:        "go.opentelemetry.io/otel/internal/global.(*nonRecordingSpan).SetStatus",
+				EntryProbe: "uprobe_SetStatus",
+				Optional:   true,
 			},
 			{
-				Sym:      "go.opentelemetry.io/otel/internal/global.(*nonRecordingSpan).SetName",
-				Fn:       uprobeSpanSetName,
-				Optional: true,
+				Sym:        "go.opentelemetry.io/otel/internal/global.(*nonRecordingSpan).SetName",
+				EntryProbe: "uprobe_SetName",
+				Optional:   true,
 			},
-		},
-
-		ReaderFn: func(obj bpfObjects) (*perf.Reader, error) {
-			return perf.NewReader(obj.Events, os.Getpagesize()*8)
 		},
 		SpecFn:    loadBpf,
 		ProcessFn: convertEvent,
 	}
 }
 
-func uprobeTracerStart(name string, exec *link.Executable, target *process.TargetDetails, obj *bpfObjects) ([]link.Link, error) {
-	offset, err := target.GetFunctionOffset(name)
-	if err != nil {
-		return nil, err
+// framePosConst is a Probe Const defining whether the tracer key contains schemaURL.
+type tracerIDContainsSchemaURL struct{}
+
+// Prior to v1.28 the tracer key did not contain schemaURL. However, in that version a
+// change was made to include it.
+// https://github.com/open-telemetry/opentelemetry-go/pull/5426/files
+var paramChangeVer = version.Must(version.NewVersion("1.28.0"))
+
+func (c tracerIDContainsSchemaURL) InjectOption(td *process.TargetDetails) (inject.Option, error) {
+	ver, ok := td.Libraries["go.opentelemetry.io/otel"]
+	if !ok {
+		return nil, fmt.Errorf("unknown module version: %s", pkg)
 	}
 
-	opts := &link.UprobeOptions{Address: offset, PID: target.PID}
-	l, err := exec.Uprobe("", obj.UprobeStart, opts)
-	if err != nil {
-		return nil, err
-	}
-
-	links := []link.Link{l}
-
-	retOffsets, err := target.GetFunctionReturns(name)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, ret := range retOffsets {
-		opts := &link.UprobeOptions{Address: ret}
-		l, err := exec.Uprobe("", obj.UprobeStartReturns, opts)
-		if err != nil {
-			return nil, err
-		}
-		links = append(links, l)
-	}
-
-	return links, nil
-}
-
-func uprobeSetAttributes(name string, exec *link.Executable, target *process.TargetDetails, obj *bpfObjects) ([]link.Link, error) {
-	offset, err := target.GetFunctionOffset(name)
-	if err != nil {
-		return nil, err
-	}
-
-	opts := &link.UprobeOptions{Address: offset, PID: target.PID}
-	l, err := exec.Uprobe("", obj.UprobeSetAttributes, opts)
-	if err != nil {
-		return nil, err
-	}
-
-	links := []link.Link{l}
-
-	return links, nil
-}
-
-func uprobeSpanSetName(name string, exec *link.Executable, target *process.TargetDetails, obj *bpfObjects) ([]link.Link, error) {
-	offset, err := target.GetFunctionOffset(name)
-	if err != nil {
-		return nil, err
-	}
-
-	opts := &link.UprobeOptions{Address: offset, PID: target.PID}
-	l, err := exec.Uprobe("", obj.UprobeSetName, opts)
-	if err != nil {
-		return nil, err
-	}
-
-	links := []link.Link{l}
-
-	return links, nil
-}
-
-func uprobeSetStatus(name string, exec *link.Executable, target *process.TargetDetails, obj *bpfObjects) ([]link.Link, error) {
-	offset, err := target.GetFunctionOffset(name)
-	if err != nil {
-		return nil, err
-	}
-
-	opts := &link.UprobeOptions{Address: offset, PID: target.PID}
-	l, err := exec.Uprobe("", obj.UprobeSetStatus, opts)
-	if err != nil {
-		return nil, err
-	}
-
-	links := []link.Link{l}
-
-	return links, nil
-}
-
-func uprobeSpanEnd(name string, exec *link.Executable, target *process.TargetDetails, obj *bpfObjects) ([]link.Link, error) {
-	offset, err := target.GetFunctionOffset(name)
-	if err != nil {
-		return nil, err
-	}
-
-	opts := &link.UprobeOptions{Address: offset, PID: target.PID}
-	l, err := exec.Uprobe("", obj.UprobeEnd, opts)
-	if err != nil {
-		return nil, err
-	}
-
-	links := []link.Link{l}
-
-	return links, nil
+	return inject.WithKeyValue("tracer_id_contains_schemaURL", ver.GreaterThanOrEqual(paramChangeVer)), nil
 }
 
 type attributeKeyVal struct {
@@ -246,12 +167,19 @@ type status struct {
 	Description [64]byte
 }
 
+type tracerID struct {
+	Name      [128]byte
+	Version   [32]byte
+	SchemaURL [128]byte
+}
+
 // event represents a manual span created by the user.
 type event struct {
 	context.BaseSpanProperties
 	SpanName   [64]byte
 	Status     status
 	Attributes attributesBuffer
+	TracerID   tracerID
 }
 
 func convertEvent(e *event) []*probe.SpanEvent {
@@ -288,6 +216,9 @@ func convertEvent(e *event) []*probe.SpanEvent {
 				Code:        codes.Code(e.Status.Code),
 				Description: string(unix.ByteSliceToString(e.Status.Description[:])),
 			},
+			TracerName:    unix.ByteSliceToString(e.TracerID.Name[:]),
+			TracerVersion: unix.ByteSliceToString(e.TracerID.Version[:]),
+			TracerSchema:  unix.ByteSliceToString(e.TracerID.SchemaURL[:]),
 		},
 	}
 }