+++ internal/pkg/instrumentation/bpf/google.golang.org/grpc/server/bpf_arm64_bpfel.go
@@ -26,8 +26,10 @@ type bpfGrpcRequestT struct {
 type bpfSliceArrayBuff struct{ Buff [1024]uint8 }
 
 type bpfSpanContext struct {
-	TraceID [16]uint8
-	SpanID  [8]uint8
+	TraceID    [16]uint8
+	SpanID     [8]uint8
+	TraceFlags uint8
+	Padding    [7]uint8
 }
 
 // loadBpf returns the embedded CollectionSpec for bpf.
@@ -83,6 +85,7 @@ type bpfMapSpecs struct {
 	AllocMap             *ebpf.MapSpec `ebpf:"alloc_map"`
 	Events               *ebpf.MapSpec `ebpf:"events"`
 	GrpcEvents           *ebpf.MapSpec `ebpf:"grpc_events"`
+	GrpcStorageMap       *ebpf.MapSpec `ebpf:"grpc_storage_map"`
 	SliceArrayBuffMap    *ebpf.MapSpec `ebpf:"slice_array_buff_map"`
 	StreamidToGrpcEvents *ebpf.MapSpec `ebpf:"streamid_to_grpc_events"`
 	TrackedSpans         *ebpf.MapSpec `ebpf:"tracked_spans"`
@@ -111,6 +114,7 @@ type bpfMaps struct {
 	AllocMap             *ebpf.Map `ebpf:"alloc_map"`
 	Events               *ebpf.Map `ebpf:"events"`
 	GrpcEvents           *ebpf.Map `ebpf:"grpc_events"`
+	GrpcStorageMap       *ebpf.Map `ebpf:"grpc_storage_map"`
 	SliceArrayBuffMap    *ebpf.Map `ebpf:"slice_array_buff_map"`
 	StreamidToGrpcEvents *ebpf.Map `ebpf:"streamid_to_grpc_events"`
 	TrackedSpans         *ebpf.Map `ebpf:"tracked_spans"`
@@ -122,6 +126,7 @@ func (m *bpfMaps) Close() error {
 		m.AllocMap,
 		m.Events,
 		m.GrpcEvents,
+		m.GrpcStorageMap,
 		m.SliceArrayBuffMap,
 		m.StreamidToGrpcEvents,
 		m.TrackedSpans,