+++ internal/pkg/instrumentation/bpf/net/http/client/probe_test.go
@@ -1,16 +1,5 @@
 // Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// SPDX-License-Identifier: Apache-2.0
 
 package client
 
@@ -21,7 +10,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
-	semconv "go.opentelemetry.io/otel/semconv/v1.25.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.26.0"
 	"go.opentelemetry.io/otel/trace"
 
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/context"
@@ -117,6 +106,7 @@ func TestConvertEvent(t *testing.T) {
 						semconv.ServerAddress(hostString),
 						semconv.NetworkProtocolVersion("1.1"),
 					},
+					TracerSchema: semconv.SchemaURL,
 				},
 			},
 		},
@@ -149,7 +139,8 @@ func TestConvertEvent(t *testing.T) {
 						semconv.ServerAddress(hostString),
 						semconv.NetworkProtocolVersion("1.1"),
 					},
-					Status: probe.Status{Code: codes.Error},
+					Status:       probe.Status{Code: codes.Error},
+					TracerSchema: semconv.SchemaURL,
 				},
 			},
 		},
@@ -182,7 +173,8 @@ func TestConvertEvent(t *testing.T) {
 						semconv.ServerAddress(hostString),
 						semconv.NetworkProtocolVersion("1.1"),
 					},
-					Status: probe.Status{Code: codes.Error},
+					Status:       probe.Status{Code: codes.Error},
+					TracerSchema: semconv.SchemaURL,
 				},
 			},
 		},
@@ -216,6 +208,7 @@ func TestConvertEvent(t *testing.T) {
 						semconv.NetworkProtocolName("foo"),
 						semconv.NetworkProtocolVersion("2.2"),
 					},
+					TracerSchema: semconv.SchemaURL,
 				},
 			},
 		},
@@ -251,6 +244,7 @@ func TestConvertEvent(t *testing.T) {
 						semconv.ServerAddress(hostString),
 						semconv.NetworkProtocolVersion("1.1"),
 					},
+					TracerSchema: semconv.SchemaURL,
 				},
 			},
 		},
@@ -286,6 +280,7 @@ func TestConvertEvent(t *testing.T) {
 						semconv.URLFull("http:/home?"),
 						semconv.NetworkProtocolVersion("1.1"),
 					},
+					TracerSchema: semconv.SchemaURL,
 				},
 			},
 		},