+++ internal/pkg/instrumentation/bpf/net/http/server/probe_test.go
@@ -1,16 +1,5 @@
 // Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// SPDX-License-Identifier: Apache-2.0
 
 package server
 
@@ -22,7 +11,7 @@ import (
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/codes"
-	semconv "go.opentelemetry.io/otel/semconv/v1.25.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.26.0"
 	"go.opentelemetry.io/otel/trace"
 
 	"go.opentelemetry.io/auto/internal/pkg/instrumentation/context"
@@ -82,6 +71,7 @@ func TestProbeConvertEvent(t *testing.T) {
 						semconv.ServerPort(8080),
 						semconv.NetworkProtocolVersion("1.1"),
 					},
+					TracerSchema: semconv.SchemaURL,
 				},
 			},
 		},
@@ -122,6 +112,7 @@ func TestProbeConvertEvent(t *testing.T) {
 						semconv.NetworkProtocolName("FOO"),
 						semconv.NetworkProtocolVersion("2.2"),
 					},
+					TracerSchema: semconv.SchemaURL,
 				},
 			},
 		},
@@ -161,6 +152,7 @@ func TestProbeConvertEvent(t *testing.T) {
 						semconv.ServerPort(8080),
 						semconv.NetworkProtocolVersion("1.1"),
 					},
+					TracerSchema: semconv.SchemaURL,
 				},
 			},
 		},
@@ -200,7 +192,8 @@ func TestProbeConvertEvent(t *testing.T) {
 						semconv.ServerPort(8080),
 						semconv.NetworkProtocolVersion("1.1"),
 					},
-					Status: probe.Status{Code: codes.Error},
+					Status:       probe.Status{Code: codes.Error},
+					TracerSchema: semconv.SchemaURL,
 				},
 			},
 		},