+++ internal/tools/inspect/inspector.go
@@ -1,16 +1,5 @@
 // Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// SPDX-License-Identifier: Apache-2.0
 
 package inspect
 
@@ -87,7 +76,7 @@ func (i *Inspector) AddManifest(manifest Manifest) error {
 
 	goVer := manifest.Application.GoVerions
 	if goVer == nil {
-		// Passsing nil to newBuilder will mean the application is built with
+		// Passing nil to newBuilder will mean the application is built with
 		// the latest version of Go.
 		b := newBuilder(i.log, i.client, nil)
 		for _, ver := range manifest.Application.Versions {